+++ src/Serializer/src/Exception/InvalidArgumentException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Spiral\Serializer\Exception;
+
+final class InvalidArgumentException extends \InvalidArgumentException implements SerializerExceptionInterface
+{
+}