+++ src/Serializer/src/Exception/SerializerExceptionInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Spiral\Serializer\Exception;
+
+interface SerializerExceptionInterface extends \Throwable
+{
+}