+++ src/Serializer/src/Exception/UnserializeException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Spiral\Serializer\Exception;
+
+final class UnserializeException extends \InvalidArgumentException implements SerializerExceptionInterface
+{
+}